# 2-2-t/specs/001-smartbooking-scheduling/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: SmartBooking API
  version: 0.1.0
  description: |
    Contract-first REST API for SmartBooking.

    Notes:
    - Auth: JWT Bearer token in Authorization header.
    - All errors use a unified envelope with stable error codes and requestId.
servers:
  - url: http://localhost:3001
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new account (USER or PROVIDER)
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (stateless token; client drops token)
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/provider/services:
    get:
      tags: [Provider]
      summary: List services owned by current provider
      operationId: providerListServices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Provider]
      summary: Create a service
      operationId: providerCreateService
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/provider/services/{serviceId}:
    patch:
      tags: [Provider]
      summary: Update a service
      operationId: providerUpdateService
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateServiceRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/provider/services/{serviceId}/timeslots:
    get:
      tags: [Provider]
      summary: List timeslots for a service
      operationId: providerListTimeSlots
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serviceId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/TimeSlot'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Provider]
      summary: Create a timeslot for a service
      operationId: providerCreateTimeSlot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/serviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTimeSlotRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/provider/timeslots/{timeSlotId}:
    patch:
      tags: [Provider]
      summary: Update a timeslot
      operationId: providerUpdateTimeSlot
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/timeSlotId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTimeSlotRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimeSlot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/services/{serviceId}:
    get:
      tags: [Users]
      summary: Get public service detail and its timeslots
      operationId: getPublicServiceDetail
      parameters:
        - $ref: '#/components/parameters/serviceId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List users
      operationId: adminListUsers
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/services:
    get:
      tags: [Admin]
      summary: List services (admin)
      operationId: adminListServices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Service'
                required: [items]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/reports/summary:
    get:
      tags: [Admin]
      summary: Report summary
      operationId: adminReportSummary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    serviceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    timeSlotId:
      name: timeSlotId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [USER, PROVIDER]
      required: [email, password, role]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        user:
          $ref: '#/components/schemas/MeResponse'
      required: [accessToken, user]
    MeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, PROVIDER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
      required: [id, email, role, status]
    CreateServiceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
      required: [name, description, durationMinutes]
    UpdateServiceRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
      additionalProperties: false
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        providerId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time
      required: [id, providerId, name, description, durationMinutes, status, createdAt]
    CreateTimeSlotRequest:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
          minimum: 0
        cancelDeadlineAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [OPEN, CLOSED]
      required: [startTime, endTime, capacity, cancelDeadlineAt]
    UpdateTimeSlotRequest:
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
          minimum: 0
        cancelDeadlineAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [OPEN, CLOSED]
      additionalProperties: false
    TimeSlot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        serviceId:
          type: string
          format: uuid
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        capacity:
          type: integer
        bookedCount:
          type: integer
        remaining:
          type: integer
          description: capacity - bookedCount
        cancelDeadlineAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [OPEN, CLOSED]
      required:
        - id
        - serviceId
        - startTime
        - endTime
        - capacity
        - bookedCount
        - remaining
        - cancelDeadlineAt
        - status
    ServiceDetailResponse:
      type: object
      properties:
        service:
          $ref: '#/components/schemas/Service'
        timeSlots:
          type: array
          items:
            $ref: '#/components/schemas/TimeSlot'
      required: [service, timeSlots]
...
    ReportSummary:
      type: object
      properties:
        totals:
          type: object
          properties:
            bookings:
              type: integer
            cancelled:
              type: integer
          required: [bookings, cancelled]
        cancellationRate:
          type: number
        activeServices:
          type: integer
        generatedAt:
          type: string
          format: date-time
      required: [totals, cancellationRate, activeServices, generatedAt]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [USER, PROVIDER, ADMIN]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        createdAt:
          type: string
          format: date-time
      required: [id, email, role, status, createdAt]
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable internal error code
              example: BOOKING_SEAT_FULL
            message:
              type: string
              description: Human-readable error message
            requestId:
              type: string
              description: Correlation id for support/debugging
          required: [code, message, requestId]
      required: [error]